Fix race condition in ICMP pinger
We currently record an inflight packet *after* sending it out. This can
lead to erroneously reporting a lost packet if the network connection is
very fast and the sending go routine is interrupted between sending the
packet and recoding that it is inflight. The receiver will drop the
returning packet if it was not recorded as inflight.

This did not show up in testing as the overlay network on the test
cluster seems to be "slow enough" to not run into this race condition.