Reduce chattiness of network-canary process
We currently produce large amounts of logs like

```
1.6825706203312843e+09 | info | ping/manager.go:105 | Updating ping targets | {"version": "0.1.1"}
```

This line is logged regardless of whether there's actually any change in
ping targets each time a round of pings is scheduled. This leads to
quite significant amount of logs which serve very little purpose, since
they're not particularly useful for debugging.

This commit changes the corresponding log statement to use `.V(1)` to
disable this log message in normal operation.